+++ src/test/java/ru/yandex/practicum/filmorate/storage/director/DbDbdbDirectorStorageTest.java
@@ -23,7 +23,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 @AutoConfigureTestDatabase
 @RequiredArgsConstructor(onConstructor_ = @Autowired)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
-public class DbDirectorStorageTest {
+public class DbDbdbDirectorStorageTest {
     @Autowired
     private FilmController filmController;
     @Autowired
